+++ packages/lumi-survey/src/global.d.ts
@@ -1,8 +1,3 @@
-declare module "*.module.css" {
-  const classes: Record<string, string>;
-  export default classes;
-}
-
 declare module "*.css" {
   const css: string;
   export default css;